+++ scripts/test.sh
@@ -92,7 +92,8 @@ run_test 'interface_vip' 1
 run_test 'interface_lagg' 1
 run_test 'interface_loopback' 1
 run_test 'interface_gre' 1
-run_test 'source_nat' 1
+run_test 'nat_source' 1
+run_test 'nat_one_to_one' 1
 run_test 'frr_diagnostic' 1
 run_test 'frr_general' 1
 run_test 'frr_bfd_general' 1
